Improve Cerata type system
* Redesigned type system to allow literal, parameter and expression nodes to be generics for types.
* Reduced verbosity of shared object construction functions.
* Removed explicit Stream type and made it a sugaring of Record.
* Added tests and fixed numerous nasty bugs.
* Follow file naming convention of vhdmmio (e.g. use .gen.vhd for generated files.)

+++ codegen/cpp/cerata/src/cerata/pool.h
@@ -1,4 +1,4 @@
-// Copyright 2018 Delft University of Technology
+// Copyright 2018-2019 Delft University of Technology
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -137,19 +137,21 @@ inline NodePool *default_node_pool() {
 // Convenience functions for fast access to literals in the default node pool.
 
 /// @brief Obtain a raw pointer to an integer literal from the default node pool.
-inline Literal *rintl(int i) {
+inline Literal *rintl(int64_t i) {
   return default_node_pool()->GetLiteral(i).get();
 }
 /// @brief Obtain a shared pointer to an integer literal from the default node pool.
-inline std::shared_ptr<Literal> intl(int i) {
+inline std::shared_ptr<Literal> intl(int64_t i) {
   return default_node_pool()->GetLiteral(i);
 }
+
 /// @brief Obtain a raw pointer to a string literal from the default node pool.
 inline Literal *rstrl(std::string str) { return default_node_pool()->GetLiteral<std::string>(std::move(str)).get(); }
 /// @brief Obtain a shared pointer to a string literal from the default node pool.
 inline std::shared_ptr<Literal> strl(std::string str) {
   return default_node_pool()->GetLiteral<std::string>(std::move(str));
 }
+
 /// @brief Return a literal node representing a Boolean.
 inline std::shared_ptr<Literal> booll(bool value) {
   return default_node_pool()->GetLiteral<bool>(value);
